Working DEB package creator

+++ selfupdate/makeDebPkg.sh
@@ -1,33 +1,61 @@
 #!/bin/bash
-#set -x # echo on
+# Run this script on the rPI with stratux cloned in ~/stratux. Make sure to have the dependencies installed:
+#   apt-get install -y dh-make libusb-1.0-0-dev mercurial wiringpi
 
-#apt-get install -y dh-make
+#set -x # echo on
 
 stratuxVersion=`git describe --tags --abbrev=0 | sed -e "s/^v//"`
 stratuxBuild=`git log -n 1 --pretty=%H`
 
 echo "Packaging ${stratuxVersion} (${stratuxBuild})."
 
+export GOROOT=/usr/local/go
+export PATH="$PATH:$GOROOT/bin"
+if [ ! -e "${GOROOT}" ]; then
+	# install Go: we need minimum 1.5 (because golang.org/x/net/internal/socket doesn't compile below 1.4)
+	wget https://storage.googleapis.com/golang/go1.6.2.linux-armv6l.tar.gz 
+	sudo tar -xzf go1.6.2.linux-armv6l.tar.gz -C /usr/local
+	rm -f go1.6.2.linux-armv6l.tar.gz
+	sudo chgrp -R staff /usr/local/go
+fi
+
+if [ "${GOPATH}" == "" ]; then
+	export GOPATH=$HOME/gocode
+	mkdir -p ${GOPATH}
+fi
+
 cd ..
-make
+sudo service stratux stop
+
+# enable multiple compiles at once
+export NUMCPU=`grep -c ^processor /proc/cpuinfo`
+make -j ${NUMCPU}
+if [[ $? -ne 0 ]]; then
+	echo "Didn't make a clean build"
+	exit 1
+fi
+
 rm -rf work
 mkdir -p work
 mkdir -p work/DEBIAN
 read -d '' control <<- EOF
 	Package: stratux 
 	Version: ${stratuxVersion} 
-	Architecture: all 
+	Architecture: armhf
 	Maintainer: Chris Young<cyoung@slack> 
 	Installed-Size: 2 
-	Depends: 
+	Depends: wiringpi
 	Section: extras 
 	Priority: optional 
-	Homepage: your homepage 
-	Description: describe 
+	Homepage: https://github.com/cyoung/stratux
+	Description: Stratux controller software 
 EOF
 echo "${control}" > work/DEBIAN/control
 
 read -d '' preinst <<- EOF
+# stop running
+service stratux stop
+
 #disable serial console
 sed -i /boot/cmdline.txt -e "s/console=ttyAMA0,[0-9]\+ //"
 
@@ -53,8 +81,13 @@ echo "${preinst}" > work/DEBIAN/preinst
 chmod 775 work/DEBIAN/preinst
 
 read -d '' postinst <<- EOF
+cp -i /root/config.txt /boot/config.txt
+
 /usr/bin/fancontrol remove
 /usr/bin/fancontrol install
+
+# start stratux
+service stratux start
 EOF
 echo "${postinst}" > work/DEBIAN/postinst
 chmod 775 work/DEBIAN/postinst
@@ -94,8 +127,10 @@ ln -fs /lib/systemd/system/stratux.service work/etc/systemd/system/multi-user.ta
 #SDR Serial Script
 	cp image/sdr-tool.sh work/usr/sbin/
 
-#boot config
-	cp image/config.txt work/boot/
+#boot config (first it's put in /root, then post-inst it is optionally placed in /boot)
+	cp image/config.txt work/root/
+	chmod 755 work/root/config.txt
+	sudo chown root.root work/root/config.txt
 
 #modprobe.d blacklist
 	cp -f image/rtl-sdr-blacklist.conf work/etc/modprobe.d/
@@ -158,3 +193,5 @@ dpkg -b work
 mv work.deb stratux-${stratuxVersion}.deb
 
 rm -rf work
+
+sudo service stratux start
